+++ src/consumer_group.rs
@@ -10,7 +10,7 @@ use crate::{
     assignor::{assign, ROUND_ROBIN_PROTOCOL},
     consumer::{ConsumeMessage, FetchParams, TopicPartitions},
     consumer_builder::ConsumerBuilder,
-    error::{KafkaCode, Result},
+    error::{Error, KafkaCode, Result},
     network::BrokerConnection,
     protocol::{
         self,
@@ -110,7 +110,10 @@ impl ConsumerGroup {
                         .collect();
 
                     let partition_assignments = assign(
-                        std::str::from_utf8(join.protocol_name.as_bytes()).unwrap(),
+                        std::str::from_utf8(join.protocol_name.as_bytes()).map_err(|err| {
+                            tracing::error!("Error converting from UTF8 {:?}", err);
+                            Error::DecodingUtf8Error
+                        })?,
                         assigned_topic_partitions,
                         number_of_consumers,
                     )?;
@@ -204,8 +207,14 @@ impl ConsumerGroup {
                 tokio::pin!(consumer);
 
                 loop {
-                    // Ugly unwrap here...
-                    yield consumer.next().await.unwrap();
+                    // If the consumer stream yields None...
+                    // We need to decide what to do,
+                    // For now we just ignore Nones and keep beating
+                    // (the consumer would yield None if it was done?
+                    // not really a thing in kafka, and ours doesn't ever do a None)
+                    if let Some(v) = consumer.next().await {
+                        yield v;
+                    }
 
                     tracing::info!("Member {:?} | Heartbeat", self.member_id);
                     let hb = heartbeat(