+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#58](https://github.com/CallistoLabsNYC/samsa/issues/58) ConsumerGroup cannot stream when caller only has a reference to a wrapper struct
 - [#47](https://github.com/CallistoLabsNYC/samsa/issues/47) Examples - Publisher and Consumer do not use same topic.
 - [#34](https://github.com/CallistoLabsNYC/samsa/issues/34) Add documentation and comments for TopicAssignments & TopicPartition structs
+- [#19](https://github.com/CallistoLabsNYC/samsa/issues/19) Add response stream for the producer
 
 ### Fixed
 - [#56](https://github.com/CallistoLabsNYC/samsa/issues/56) Static Analysis and Build CI actions run with no purpose during tag release